+++ source/AudioTraits-FrequencyDomain.hpp
